+++ R/dt.R
@@ -1,7 +1,7 @@
 #' Pipeable data.table call
 #'
 #' @description
-#' Pipeable data.table call
+#' Pipeable data.table call with (experimental) support for tidy evaluation.
 #'
 #' Note: This function does not use data.table's modify-by-reference
 #'
@@ -18,6 +18,15 @@
 #' df %>%
 #'   dt(, double_x := x * 2) %>%
 #'   dt(order(-double_x))
+#'
+#' # Experimental support for tidy evaluation
+#' add_one <- function(data, col) {
+#'   data %>%
+#'     dt(, {{ col }} := {{ col }} + 1)
+#' }
+#'
+#' df %>%
+#'   add_one(x)
 #' @export
 dt <- function(.df, ...) {
   UseMethod("dt")
@@ -25,9 +34,7 @@ dt <- function(.df, ...) {
 
 #' @export
 dt.tidytable <- function(.df, ...) {
-  # TODO: Add test that let() doesn't modify-by-reference
-    ## once 1.14.4 is released
-  dots <- enquos(..., .unquote_names = FALSE)
+  dots <- enquos(..., .unquote_names = FALSE, .ignore_empty = "none")
   dt_env <- get_dt_env(dots)
   dots <- lapply(dots, quo_squash)
   dots_names <- names(dots)
@@ -40,7 +47,7 @@ dt.tidytable <- function(.df, ...) {
     }
 
     if (is_call(j, c(":=", "let"))) {
-      mut_exprs <- j[-1]
+      mut_exprs <- as.list(j[-1])
       if (length(mut_exprs) == 2 && is.null(names(mut_exprs))) {
         col_name <- mut_exprs[[1]]
         if (is_call(col_name, "(")) {
@@ -57,8 +64,20 @@ dt.tidytable <- function(.df, ...) {
         .df <- fast_copy(.df, col_name)
       } else {
         # .df[, let(x = 1, double_y = y * 2)]
-        col_names <- names(mut_exprs)
+        use_walrus <- map_lgl.(mut_exprs, is_call, ":=")
+        if (any(use_walrus)) {
+          j <- prep_j_expr(mut_exprs, use_walrus, ":=")
+          dots <- replace_j_dot(dots, dots_names, j)
+        }
+        col_names <- names(as.list(j[-1]))
         .df <- fast_copy(.df, col_names)
+      }
+    } else if (is_call(j, c(".", "list"))) {
+      summarize_exprs <- as.list(j[-1])
+      use_walrus <- map_lgl.(summarize_exprs, is_call, ":=")
+      if (any(use_walrus)) {
+        j <- prep_j_expr(summarize_exprs, use_walrus, ".")
+        dots <- replace_j_dot(dots, dots_names, j)
       }
     }
   }
@@ -66,11 +85,32 @@ dt.tidytable <- function(.df, ...) {
   dt_expr <- call2("[", quo(.df), !!!dots)
 
   # Only add empty `[` when using mutate
-  if (exists("mut_exprs", envir = current_env())) {
+  if (exists("mut_exprs", current_env())) {
     dt_expr <- call2("[", dt_expr)
   }
 
   eval_tidy(dt_expr, env = dt_env)
+}
+
+prep_j_expr <- function(j_exprs, use_walrus, j_call) {
+  walrus_exprs <- j_exprs[use_walrus]
+  walrus_exprs <- map.(walrus_exprs, ~ as.list(.x[-1]))
+  walrus_names <- map_chr.(walrus_exprs, ~ as.character(.x[[1]]))
+  walrus_exprs <- map.(walrus_exprs, ~ .x[[2]])
+  j_exprs[use_walrus] <- walrus_exprs
+  names(j_exprs)[use_walrus] <- walrus_names
+
+  new_j <- call2(j_call, !!!j_exprs)
+  new_j
+}
+
+replace_j_dot <- function(dots, dots_names, j) {
+  if ("j" %chin% dots_names) {
+    dots[["j"]] <- j
+  } else {
+    dots[[2]] <- j
+  }
+  dots
 }
 
 #' @export